prep to move key endpoints to JWT auth. updates routes & handlers. 5 tests failing. dont panic! :-)
