+++ handlers/home.js
@@ -1,8 +1,6 @@
-module.exports = {
-  handler: function(req, reply) {
-    // console.log(req);
-    // need to define exactly what the API should return on home page.
-    // e.g: please register at timerzzz.com
-    reply('Welcome to Timer Land!');
-  }
+module.exports = function(req, reply) {
+  // need to define exactly what the API should return on home page.
+  // e.g: please register at timerzzz.com
+  // please make suggestions: https://github.com/ideaq/time/issues/46
+  reply('Welcome to Timer Land!');
 }
